+++ fiddles/c/fiddle-0000-Template/template/src/main.c
@@ -1,6 +1,11 @@
 #include <stdio.h>
 
 int main(void) {
-    printf ("template compiled!");
+    printf ("\ntemplate started\n\n");
+
+    // Do stuff
+
+    printf ("template ended\n\n");
+
     return 0;
 }