+++ fiddles/c/template/main.c
@@ -1,6 +1,11 @@
 #include <stdio.h>
 
 int main(void) {
-    printf ("{{projectName}} compiled!");
+    printf ("\n{{projectName}} started\n\n");
+
+    // Do stuff
+
+    printf ("{{projectName}} ended\n\n");
+
     return 0;
 }